+++ HaloEzAPI.Tests/HaloAPIServiceTests/MetaData/HaloWars2/GetGameObjectsTests.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Linq;
+using NUnit.Framework;
+
+namespace HaloEzAPI.Tests.HaloAPIServiceTests.MetaData.HaloWars2
+{
+    [TestFixture]
+    public class GetGameObjectsTests : BaseHaloAPIServiceTests
+    {
+        [Test]
+        public void Default_DoesNotThrowException()
+        {
+            Assert.DoesNotThrow(async () => await HaloApiService.HaloWars2.GetGameObjects());
+        }
+
+        [Test]
+        public async void Default_DoesNotReturnNull()
+        {
+            var result = await HaloApiService.HaloWars2.GetGameObjects();
+            Assert.IsNotNull(result);
+        }
+
+        [Test]
+        public async void Default_ReturnsValidCollection()
+        {
+            var result = await HaloApiService.HaloWars2.GetGameObjects();
+            CollectionAssert.AllItemsAreNotNull(result.ContentItems);
+        }
+
+        [Test]
+        public async void Default_ReturnsKnownGameObjects()
+        {
+            var result = await HaloApiService.HaloWars2.GetGameObjects();
+            var firstGameObject = result.ContentItems.First();
+            
+            Assert.AreEqual(firstGameObject.View.Title, "VOLATILE SCARAB");
+        }
+    }
+}