+++ HaloEzAPI/Abstraction/Enum/HaloWars2/EffectivenessAgainstGameObject.cs
@@ -1,0 +1,10 @@
+﻿namespace HaloEzAPI.Abstraction.Enum.HaloWars2
+{
+    public enum EffectivenessAgainstGameObject
+    {
+        NotApplicable,
+        Poor,
+        Neutral,
+        Good,
+    }
+}