+++ HaloEzAPI/Model/Response/MetaData/HaloWars2/Cards/HW2LeaderPowerInnerViewDisplayInfo.cs
@@ -1,0 +1,9 @@
+using HaloEzAPI.Model.Response.MetaData.HaloWars2.Views;
+
+namespace HaloEzAPI.Model.Response.MetaData.HaloWars2.Cards
+{
+    public class HW2LeaderPowerInnerViewDisplayInfo : BaseView
+    {
+        public HW2LeaderPowerDisplayInfo HW2LeaderPowerDisplayInfo { get; set; }
+    }
+}