+++ HaloEzAPI/Model/Response/MetaData/HaloWars2/Cards/HW2StartingArmyDisplayInfo.cs
@@ -2,7 +2,7 @@ using HaloEzAPI.Model.Response.MetaData.HaloWars2.Shared;
 
 namespace HaloEzAPI.Model.Response.MetaData.HaloWars2.Cards
 {
-    public class HW2DifficultyDisplayInfo: Details
+    public class HW2StartingArmyDisplayInfo : Details
     {
     }
 }