+++ HaloEzAPI/Model/Response/MetaData/HaloWars2/Views/HW2LeaderPowerView.cs
@@ -1,0 +1,9 @@
+﻿using HaloEzAPI.Model.Response.MetaData.HaloWars2.Cards;
+
+namespace HaloEzAPI.Model.Response.MetaData.HaloWars2.Views
+{
+    public class HW2LeaderPowerView : BaseView
+    {
+        public HW2LeaderPower HW2LeaderPower { get; set; }
+    }
+}