+++ HaloEzAPI/Model/Response/MetaData/HaloWars2/Views/HW2PlaylistView.cs
@@ -1,0 +1,9 @@
+﻿using HaloEzAPI.Model.Response.MetaData.HaloWars2.Cards;
+
+namespace HaloEzAPI.Model.Response.MetaData.HaloWars2.Views
+{
+    public class HW2PlaylistView : BaseView
+    {
+        public HW2Playlist HW2Playlist { get; set; }
+    }
+}