+++ HaloEzAPI/Model/Response/Stats/Halo5/ProgressiveCommendationDelta.cs
@@ -5,7 +5,7 @@ namespace HaloEzAPI.Model.Response.Stats.Halo5
     public class ProgressiveCommendationDelta
     {
         public Guid Id { get; set; }
-        public int PreviousPrOgress { get; set; }
+        public int PreviousProgress { get; set; }
         public int Progress { get; set; }
     }
 }